refactor(phone-sensors): android specific code in android files
